api/requirements.txt
Django==3.1.4
django-allauth==0.44.0
django-anymail==8.2
djangorestframework==3.12.2
mysqlclient==2.0.2
python-dotenv==0.15.0
